+++ page_object/pages/address.py
@@ -1,12 +1,12 @@
 """address page"""
 
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import TimeoutException, NoSuchElementException, \
+from selenium.common.exceptions import NoSuchElementException, \
     InvalidSelectorException
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 
-from test.page_object.locators import Locator
+from page_object.locators import Locator
 
 
 class Address:
