+++ page_object/pages/authentication.py
@@ -1,9 +1,9 @@
 """Authentication page"""
 
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import TimeoutException, NoSuchElementException, InvalidSelectorException
+from selenium.common.exceptions import NoSuchElementException, InvalidSelectorException
 
-from test.page_object.locators import Locator
+from page_object.locators import Locator
 
 
 class Authentication:
