+++ page_object/pages/home.py
@@ -1,9 +1,9 @@
 """home page"""
 
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import TimeoutException, NoSuchElementException, InvalidSelectorException
+from selenium.common.exceptions import NoSuchElementException, InvalidSelectorException
 
-from test.page_object.locators import Locator
+from page_object.locators import Locator
 
 
 class Home:
