+++ page_object/pages/myaccount.py
@@ -3,9 +3,9 @@
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-from selenium.common.exceptions import TimeoutException, NoSuchElementException, InvalidSelectorException
+from selenium.common.exceptions import NoSuchElementException, InvalidSelectorException
 
-from test.page_object.locators import Locator
+from page_object.locators import Locator
 
 
 class MyAccount:
